IndexedDB operations in a Page fail after a StorageProcess crash. <rdar://problem/41626526> and https://bugs.webkit.org/show_bug.cgi?id=187123
Reviewed by Alex Christensen.

Source/WebCore:

Test: storage/indexeddb/modern/opendatabase-after-storage-crash.html

When the connection to a StorageProcess goes away, explicitly tell all of the WebPages
in the WebProcess about it.

This puts Documents/Workers in an error mode where requests fail instead of timeout.
It also clears the Page's connection so *new* Documents and Workers will get a fresh
new connection that works.

* Modules/indexeddb/client/IDBConnectionToServer.cpp:
(WebCore::IDBClient::IDBConnectionToServer::callResultFunctionLater):
(WebCore::IDBClient::IDBConnectionToServer::deleteDatabase):
(WebCore::IDBClient::IDBConnectionToServer::openDatabase):
(WebCore::IDBClient::IDBConnectionToServer::createObjectStore):
(WebCore::IDBClient::IDBConnectionToServer::deleteObjectStore):
(WebCore::IDBClient::IDBConnectionToServer::renameObjectStore):
(WebCore::IDBClient::IDBConnectionToServer::clearObjectStore):
(WebCore::IDBClient::IDBConnectionToServer::createIndex):
(WebCore::IDBClient::IDBConnectionToServer::deleteIndex):
(WebCore::IDBClient::IDBConnectionToServer::renameIndex):
(WebCore::IDBClient::IDBConnectionToServer::putOrAdd):
(WebCore::IDBClient::IDBConnectionToServer::getRecord):
(WebCore::IDBClient::IDBConnectionToServer::getAllRecords):
(WebCore::IDBClient::IDBConnectionToServer::getCount):
(WebCore::IDBClient::IDBConnectionToServer::deleteRecord):
(WebCore::IDBClient::IDBConnectionToServer::openCursor):
(WebCore::IDBClient::IDBConnectionToServer::iterateCursor):
(WebCore::IDBClient::IDBConnectionToServer::establishTransaction):
(WebCore::IDBClient::IDBConnectionToServer::commitTransaction):
(WebCore::IDBClient::IDBConnectionToServer::didFinishHandlingVersionChangeTransaction):
(WebCore::IDBClient::IDBConnectionToServer::abortTransaction):
(WebCore::IDBClient::IDBConnectionToServer::didFireVersionChangeEvent):
(WebCore::IDBClient::IDBConnectionToServer::confirmDidCloseFromServer):
(WebCore::IDBClient::IDBConnectionToServer::connectionToServerLost):
(WebCore::IDBClient::IDBConnectionToServer::openDBRequestCancelled):
(WebCore::IDBClient::IDBConnectionToServer::databaseConnectionPendingClose):
(WebCore::IDBClient::IDBConnectionToServer::databaseConnectionClosed):
(WebCore::IDBClient::IDBConnectionToServer::abortOpenAndUpgradeNeeded):
(WebCore::IDBClient::IDBConnectionToServer::getAllDatabaseNames):
* Modules/indexeddb/client/IDBConnectionToServer.h:

* Modules/indexeddb/shared/IDBError.h:
(WebCore::IDBError::serverConnectionLostError):

* page/Page.cpp:
(WebCore::Page::setSessionID):
(WebCore::Page::idbConnection):
(WebCore::Page::optionalIDBConnection):
(WebCore::Page::clearIDBConnection):
* page/Page.h:

Source/WebKit:

When the connection to a StorageProcess goes away, explicitly tell all of the WebPages
in the WebProcess about it.

This puts Documents/Workers in an error mode where requests fail instead of timeout.
It also clears the Page's connection so *new* Documents and Workers will get a fresh
new connection that works.

* UIProcess/API/C/WKContext.cpp:
(WKContextTerminateStorageProcess):

* UIProcess/API/Cocoa/WKProcessPool.mm:
(-[WKProcessPool _terminateStorageProcess]):

* UIProcess/Storage/StorageProcessProxy.cpp:
(WebKit::StorageProcessProxy::terminateForTesting):
* UIProcess/Storage/StorageProcessProxy.h:

* UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::terminateStorageProcessForTesting):
(WebKit::WebProcessPool::terminateStorageProcess): Deleted.
* UIProcess/WebProcessPool.h:

* WebProcess/Storage/WebToStorageProcessConnection.cpp:
(WebKit::WebToStorageProcessConnection::didClose):
* WebProcess/Storage/WebToStorageProcessConnection.h:
(WebKit::WebToStorageProcessConnection::existingIDBConnectionToServerForIdentifier):

* WebProcess/WebProcess.cpp:
(WebKit::WebProcess::webToStorageProcessConnectionClosed):

LayoutTests:

* storage/indexeddb/modern/opendatabase-after-storage-crash.html: Added.


Canonical link: https://commits.webkit.org/202639@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233562 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/Modules/indexeddb/client/IDBConnectionToServer.cpp
@@ -64,10 +64,21 @@ IDBConnectionProxy& IDBConnectionToServer::proxy()
     return *m_proxy;
 }
 
+void IDBConnectionToServer::callResultFunctionWithErrorLater(ResultFunction function, const IDBResourceIdentifier& requestIdentifier)
+{
+    callOnMainThread([this, protectedThis = makeRef(*this), function, requestIdentifier]() {
+        (this->*function)(IDBResultData::error(requestIdentifier, IDBError::serverConnectionLostError()));
+    });
+}
+
 void IDBConnectionToServer::deleteDatabase(const IDBRequestData& request)
 {
     LOG(IndexedDB, "IDBConnectionToServer::deleteDatabase - %s", request.databaseIdentifier().debugString().utf8().data());
-    m_delegate->deleteDatabase(request);
+    
+    if (m_serverConnectionIsValid)
+        m_delegate->deleteDatabase(request);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didDeleteDatabase, request.requestIdentifier());
 }
 
 void IDBConnectionToServer::didDeleteDatabase(const IDBResultData& resultData)
@@ -79,7 +90,11 @@ void IDBConnectionToServer::didDeleteDatabase(const IDBResultData& resultData)
 void IDBConnectionToServer::openDatabase(const IDBRequestData& request)
 {
     LOG(IndexedDB, "IDBConnectionToServer::openDatabase - %s (%s) (%" PRIu64 ")", request.databaseIdentifier().debugString().utf8().data(), request.requestIdentifier().loggingString().utf8().data(), request.requestedVersion());
-    m_delegate->openDatabase(request);
+
+    if (m_serverConnectionIsValid)
+        m_delegate->openDatabase(request);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didOpenDatabase, request.requestIdentifier());
 }
 
 void IDBConnectionToServer::didOpenDatabase(const IDBResultData& resultData)
@@ -93,7 +108,10 @@ void IDBConnectionToServer::createObjectStore(const IDBRequestData& requestData,
     LOG(IndexedDB, "IDBConnectionToServer::createObjectStore");
     ASSERT(isMainThread());
 
-    m_delegate->createObjectStore(requestData, info);
+    if (m_serverConnectionIsValid)
+        m_delegate->createObjectStore(requestData, info);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didCreateObjectStore, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didCreateObjectStore(const IDBResultData& resultData)
@@ -107,7 +125,10 @@ void IDBConnectionToServer::deleteObjectStore(const IDBRequestData& requestData,
     LOG(IndexedDB, "IDBConnectionToServer::deleteObjectStore");
     ASSERT(isMainThread());
 
-    m_delegate->deleteObjectStore(requestData, objectStoreName);
+    if (m_serverConnectionIsValid)
+        m_delegate->deleteObjectStore(requestData, objectStoreName);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didDeleteObjectStore, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didDeleteObjectStore(const IDBResultData& resultData)
@@ -121,7 +142,10 @@ void IDBConnectionToServer::renameObjectStore(const IDBRequestData& requestData,
     LOG(IndexedDB, "IDBConnectionToServer::renameObjectStore");
     ASSERT(isMainThread());
 
-    m_delegate->renameObjectStore(requestData, objectStoreIdentifier, newName);
+    if (m_serverConnectionIsValid)
+        m_delegate->renameObjectStore(requestData, objectStoreIdentifier, newName);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didRenameObjectStore, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didRenameObjectStore(const IDBResultData& resultData)
@@ -135,7 +159,10 @@ void IDBConnectionToServer::clearObjectStore(const IDBRequestData& requestData, 
     LOG(IndexedDB, "IDBConnectionToServer::clearObjectStore");
     ASSERT(isMainThread());
 
-    m_delegate->clearObjectStore(requestData, objectStoreIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->clearObjectStore(requestData, objectStoreIdentifier);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didClearObjectStore, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didClearObjectStore(const IDBResultData& resultData)
@@ -149,7 +176,10 @@ void IDBConnectionToServer::createIndex(const IDBRequestData& requestData, const
     LOG(IndexedDB, "IDBConnectionToServer::createIndex");
     ASSERT(isMainThread());
 
-    m_delegate->createIndex(requestData, info);
+    if (m_serverConnectionIsValid)
+        m_delegate->createIndex(requestData, info);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didCreateIndex, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didCreateIndex(const IDBResultData& resultData)
@@ -163,7 +193,10 @@ void IDBConnectionToServer::deleteIndex(const IDBRequestData& requestData, uint6
     LOG(IndexedDB, "IDBConnectionToServer::deleteIndex");
     ASSERT(isMainThread());
 
-    m_delegate->deleteIndex(requestData, objectStoreIdentifier, indexName);
+    if (m_serverConnectionIsValid)
+        m_delegate->deleteIndex(requestData, objectStoreIdentifier, indexName);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didDeleteIndex, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didDeleteIndex(const IDBResultData& resultData)
@@ -177,7 +210,10 @@ void IDBConnectionToServer::renameIndex(const IDBRequestData& requestData, uint6
     LOG(IndexedDB, "IDBConnectionToServer::renameIndex");
     ASSERT(isMainThread());
 
-    m_delegate->renameIndex(requestData, objectStoreIdentifier, indexIdentifier, newName);
+    if (m_serverConnectionIsValid)
+        m_delegate->renameIndex(requestData, objectStoreIdentifier, indexIdentifier, newName);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didRenameIndex, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didRenameIndex(const IDBResultData& resultData)
@@ -191,7 +227,10 @@ void IDBConnectionToServer::putOrAdd(const IDBRequestData& requestData, const ID
     LOG(IndexedDB, "IDBConnectionToServer::putOrAdd");
     ASSERT(isMainThread());
 
-    m_delegate->putOrAdd(requestData, key, value, overwriteMode);
+    if (m_serverConnectionIsValid)
+        m_delegate->putOrAdd(requestData, key, value, overwriteMode);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didPutOrAdd, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didPutOrAdd(const IDBResultData& resultData)
@@ -206,7 +245,10 @@ void IDBConnectionToServer::getRecord(const IDBRequestData& requestData, const I
     ASSERT(isMainThread());
     ASSERT(!getRecordData.keyRangeData.isNull);
 
-    m_delegate->getRecord(requestData, getRecordData);
+    if (m_serverConnectionIsValid)
+        m_delegate->getRecord(requestData, getRecordData);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didGetRecord, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didGetRecord(const IDBResultData& resultData)
@@ -220,7 +262,10 @@ void IDBConnectionToServer::getAllRecords(const IDBRequestData& requestData, con
     LOG(IndexedDB, "IDBConnectionToServer::getAllRecords");
     ASSERT(isMainThread());
 
-    m_delegate->getAllRecords(requestData, getAllRecordsData);
+    if (m_serverConnectionIsValid)
+        m_delegate->getAllRecords(requestData, getAllRecordsData);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didGetAllRecords, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didGetAllRecords(const IDBResultData& resultData)
@@ -235,7 +280,10 @@ void IDBConnectionToServer::getCount(const IDBRequestData& requestData, const ID
     ASSERT(isMainThread());
     ASSERT(!keyRangeData.isNull);
 
-    m_delegate->getCount(requestData, keyRangeData);
+    if (m_serverConnectionIsValid)
+        m_delegate->getCount(requestData, keyRangeData);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didGetCount, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didGetCount(const IDBResultData& resultData)
@@ -250,7 +298,10 @@ void IDBConnectionToServer::deleteRecord(const IDBRequestData& requestData, cons
     ASSERT(isMainThread());
     ASSERT(!keyRangeData.isNull);
 
-    m_delegate->deleteRecord(requestData, keyRangeData);
+    if (m_serverConnectionIsValid)
+        m_delegate->deleteRecord(requestData, keyRangeData);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didDeleteRecord, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didDeleteRecord(const IDBResultData& resultData)
@@ -264,7 +315,10 @@ void IDBConnectionToServer::openCursor(const IDBRequestData& requestData, const 
     LOG(IndexedDB, "IDBConnectionToServer::openCursor");
     ASSERT(isMainThread());
 
-    m_delegate->openCursor(requestData, info);
+    if (m_serverConnectionIsValid)
+        m_delegate->openCursor(requestData, info);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didOpenCursor, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didOpenCursor(const IDBResultData& resultData)
@@ -278,7 +332,10 @@ void IDBConnectionToServer::iterateCursor(const IDBRequestData& requestData, con
     LOG(IndexedDB, "IDBConnectionToServer::iterateCursor");
     ASSERT(isMainThread());
 
-    m_delegate->iterateCursor(requestData, data);
+    if (m_serverConnectionIsValid)
+        m_delegate->iterateCursor(requestData, data);
+    else
+        callResultFunctionWithErrorLater(&IDBConnectionToServer::didIterateCursor, requestData.requestIdentifier());
 }
 
 void IDBConnectionToServer::didIterateCursor(const IDBResultData& resultData)
@@ -292,7 +349,8 @@ void IDBConnectionToServer::establishTransaction(uint64_t databaseConnectionIden
     LOG(IndexedDB, "IDBConnectionToServer::establishTransaction");
     ASSERT(isMainThread());
 
-    m_delegate->establishTransaction(databaseConnectionIdentifier, info);
+    if (m_serverConnectionIsValid)
+        m_delegate->establishTransaction(databaseConnectionIdentifier, info);
 }
 
 void IDBConnectionToServer::commitTransaction(const IDBResourceIdentifier& transactionIdentifier)
@@ -300,7 +358,13 @@ void IDBConnectionToServer::commitTransaction(const IDBResourceIdentifier& trans
     LOG(IndexedDB, "IDBConnectionToServer::commitTransaction");
     ASSERT(isMainThread());
 
-    m_delegate->commitTransaction(transactionIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->commitTransaction(transactionIdentifier);
+    else {
+        callOnMainThread([this, protectedThis = makeRef(*this), transactionIdentifier] {
+            didCommitTransaction(transactionIdentifier, IDBError::serverConnectionLostError());
+        });
+    }
 }
 
 void IDBConnectionToServer::didCommitTransaction(const IDBResourceIdentifier& transactionIdentifier, const IDBError& error)
@@ -316,7 +380,8 @@ void IDBConnectionToServer::didFinishHandlingVersionChangeTransaction(uint64_t d
     LOG(IndexedDB, "IDBConnectionToServer::didFinishHandlingVersionChangeTransaction");
     ASSERT(isMainThread());
 
-    m_delegate->didFinishHandlingVersionChangeTransaction(databaseConnectionIdentifier, transactionIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->didFinishHandlingVersionChangeTransaction(databaseConnectionIdentifier, transactionIdentifier);
 }
 
 void IDBConnectionToServer::abortTransaction(const IDBResourceIdentifier& transactionIdentifier)
@@ -324,7 +389,13 @@ void IDBConnectionToServer::abortTransaction(const IDBResourceIdentifier& transa
     LOG(IndexedDB, "IDBConnectionToServer::abortTransaction");
     ASSERT(isMainThread());
 
-    m_delegate->abortTransaction(transactionIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->abortTransaction(transactionIdentifier);
+    else {
+        callOnMainThread([this, protectedThis = makeRef(*this), transactionIdentifier] {
+            didAbortTransaction(transactionIdentifier, IDBError::serverConnectionLostError());
+        });
+    }
 }
 
 void IDBConnectionToServer::didAbortTransaction(const IDBResourceIdentifier& transactionIdentifier, const IDBError& error)
@@ -348,7 +419,8 @@ void IDBConnectionToServer::didFireVersionChangeEvent(uint64_t databaseConnectio
     LOG(IndexedDB, "IDBConnectionToServer::didFireVersionChangeEvent");
     ASSERT(isMainThread());
 
-    m_delegate->didFireVersionChangeEvent(databaseConnectionIdentifier, requestIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->didFireVersionChangeEvent(databaseConnectionIdentifier, requestIdentifier);
 }
 
 void IDBConnectionToServer::didStartTransaction(const IDBResourceIdentifier& transactionIdentifier, const IDBError& error)
@@ -372,14 +444,17 @@ void IDBConnectionToServer::confirmDidCloseFromServer(uint64_t databaseConnectio
     LOG(IndexedDB, "IDBConnectionToServer::confirmDidCloseFromServer");
     ASSERT(isMainThread());
 
-    m_delegate->confirmDidCloseFromServer(databaseConnectionIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->confirmDidCloseFromServer(databaseConnectionIdentifier);
 }
 
 void IDBConnectionToServer::connectionToServerLost(const IDBError& error)
 {
     LOG(IndexedDB, "IDBConnectionToServer::connectionToServerLost");
     ASSERT(isMainThread());
-
+    ASSERT(m_serverConnectionIsValid);
+    
+    m_serverConnectionIsValid = false;
     m_proxy->connectionToServerLost(error);
 }
 
@@ -396,7 +471,8 @@ void IDBConnectionToServer::openDBRequestCancelled(const IDBRequestData& request
     LOG(IndexedDB, "IDBConnectionToServer::openDBRequestCancelled");
     ASSERT(isMainThread());
 
-    m_delegate->openDBRequestCancelled(requestData);
+    if (m_serverConnectionIsValid)
+        m_delegate->openDBRequestCancelled(requestData);
 }
 
 void IDBConnectionToServer::databaseConnectionPendingClose(uint64_t databaseConnectionIdentifier)
@@ -404,7 +480,8 @@ void IDBConnectionToServer::databaseConnectionPendingClose(uint64_t databaseConn
     LOG(IndexedDB, "IDBConnectionToServer::databaseConnectionPendingClose");
     ASSERT(isMainThread());
 
-    m_delegate->databaseConnectionPendingClose(databaseConnectionIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->databaseConnectionPendingClose(databaseConnectionIdentifier);
 }
 
 void IDBConnectionToServer::databaseConnectionClosed(uint64_t databaseConnectionIdentifier)
@@ -412,7 +489,8 @@ void IDBConnectionToServer::databaseConnectionClosed(uint64_t databaseConnection
     LOG(IndexedDB, "IDBConnectionToServer::databaseConnectionClosed");
     ASSERT(isMainThread());
 
-    m_delegate->databaseConnectionClosed(databaseConnectionIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->databaseConnectionClosed(databaseConnectionIdentifier);
 }
 
 void IDBConnectionToServer::abortOpenAndUpgradeNeeded(uint64_t databaseConnectionIdentifier, const IDBResourceIdentifier& transactionIdentifier)
@@ -420,7 +498,8 @@ void IDBConnectionToServer::abortOpenAndUpgradeNeeded(uint64_t databaseConnectio
     LOG(IndexedDB, "IDBConnectionToServer::abortOpenAndUpgradeNeeded");
     ASSERT(isMainThread());
 
-    m_delegate->abortOpenAndUpgradeNeeded(databaseConnectionIdentifier, transactionIdentifier);
+    if (m_serverConnectionIsValid)
+        m_delegate->abortOpenAndUpgradeNeeded(databaseConnectionIdentifier, transactionIdentifier);
 }
 
 void IDBConnectionToServer::getAllDatabaseNames(const SecurityOrigin& mainFrameOrigin, const SecurityOrigin& openingOrigin, Function<void (const Vector<String>&)>&& callback)
@@ -429,7 +508,13 @@ void IDBConnectionToServer::getAllDatabaseNames(const SecurityOrigin& mainFrameO
 
     m_getAllDatabaseNamesCallbacks.add(++callbackID, WTFMove(callback));
 
-    m_delegate->getAllDatabaseNames(mainFrameOrigin.data(), openingOrigin.data(), callbackID);
+    if (m_serverConnectionIsValid)
+        m_delegate->getAllDatabaseNames(mainFrameOrigin.data(), openingOrigin.data(), callbackID);
+    else {
+        callOnMainThread([this, protectedThis = makeRef(*this), callbackID = callbackID] {
+            didGetAllDatabaseNames(callbackID, { });
+        });
+    }
 }
 
 void IDBConnectionToServer::didGetAllDatabaseNames(uint64_t callbackID, const Vector<String>& databaseNames)
